+++ core/chains/evm/headtracker/head_saver_test.go
@@ -49,7 +49,7 @@ func TestHeadSaver_Save(t *testing.T) {
 	require.Equal(t, int64(1), latest.Number)
 }
 
-func TestHeadSaver_LoadFromDB(t *testing.T) {
+func TestHeadSaver_Load(t *testing.T) {
 	t.Parallel()
 
 	saver, orm := configureSaver(t)
@@ -59,7 +59,7 @@ func TestHeadSaver_LoadFromDB(t *testing.T) {
 		require.NoError(t, err)
 	}
 
-	latestHead, err := saver.LoadFromDB(testutils.Context(t))
+	latestHead, err := saver.Load(testutils.Context(t))
 	require.NoError(t, err)
 	require.NotNil(t, latestHead)
 	require.Equal(t, int64(4), latestHead.Number)