+++ core/cmd/solana_keys_commands_test.go
@@ -100,7 +100,8 @@ func TestClient_SolanaKeys(t *testing.T) {
 		require.NoError(tt, set.Set("yes", "true"))
 
 		strID := key.ID()
-		set.Parse([]string{strID})
+		err = set.Parse([]string{strID})
+		require.NoError(t, err)
 		c := cli.NewContext(nil, set, nil)
 		err = cmd.NewSolanaKeysClient(client).DeleteKey(c)
 		require.NoError(t, err)